[CSharpBinding] Refactored presentation layer
+ Removed presentation layer from IDE - needs to move to the c# binding for now.
+ Corrected namespaces & directory of the presentation layer
+ Splitted up the presenter into a view-model
