+++ main/src/addins/CSharpBinding/CSharpBinding.csproj
@@ -309,17 +309,6 @@
     <Compile Include="MonoDevelop.CSharp.Features\Formatter\Indent.cs" />
     <Compile Include="MonoDevelop.CSharp\DeclaredSymbolInfo.cs" />
     <Compile Include="MonoDevelop.CSharp\ProjectSearchCategory.cs" />
-    <Compile Include="Razor\ContainedDocumentPreserveFormattingRule.cs" />
-    <Compile Include="Razor\IMonoDevelopContainedLanguageHost.cs" />
-    <Compile Include="Razor\IMyRoslynCompletionDataProvider.cs" />
-    <Compile Include="Razor\MonoDevelopContainedDocument.cs" />
-    <Compile Include="Razor\MonoDevelopFormattingRuleFactoryServiceFactory.cs" />
-    <Compile Include="Razor\MyCSharpCompletionData.cs" />
-    <Compile Include="Razor\MyCSharpCompletionDataProvider.cs" />
-    <Compile Include="Razor\MyRoslynCompletionData.cs" />
-    <Compile Include="Razor\RoslynCommandTarget.cs" />
-    <Compile Include="Razor\RoslynCompletionPresenter.cs" />
-    <Compile Include="Razor\RoslynCompletionPresenterSession.cs" />
     <Compile Include="Util\7BitEncodedInts.cs" />
     <Compile Include="Util\ArgumentSyntaxExtensions.cs" />
     <Compile Include="Util\CloneableStack.cs" />
@@ -378,6 +367,19 @@
     <Compile Include="MonoDevelop.CSharp.Navigation\Counters.cs" />
     <Compile Include="MonoDevelop.CSharp.Project\CSharpLanguageVersionHelper.cs" />
     <Compile Include="MonoDevelop.CSharp.Highlighting\HighlightUsagesExtension.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\RoslynCompletionPresenter.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\RoslynCompletionPresenterSession.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\ContainedDocumentPreserveFormattingRule.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\IMonoDevelopContainedLanguageHost.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\IMyRoslynCompletionDataProvider.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\MonoDevelopContainedDocument.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\MonoDevelopFormattingRuleFactoryServiceFactory.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\MyCSharpCompletionData.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\MyCSharpCompletionDataProvider.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\MyRoslynCompletionData.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\RoslynCommandTarget.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\GtkCompletionPresenterView.cs" />
+    <Compile Include="MonoDevelop.CSharp.Completion.Presentation\ICompletionPresenterView.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Makefile.am" />
@@ -396,6 +398,9 @@
       <PublicKey>0024000004800000940000000602000000240000525341310004000001000100b5fc90e7027f67871e773a8fde8938c81dd402ba65b9201d60593e96c492651e889cc13f1415ebb53fac1131ae0bd333c5ee6021672d9718ea31a8aebd0da0072f25d87dba6fc90ffd598ed4da35e44c398c454307e8e33b8426143daec9f596836f97c8f74750e5975c64e2189f45def46b2a2b1247adc3652bf5c308055da9</PublicKey>
     </InternalsVisibleTo>
     <InternalsVisibleTo Include="WebToolingAddin" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="MonoDevelop.CSharp.Completion.Presentation\" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
  </Project>
