+++ main/src/addins/CSharpBinding/MonoDevelop.CSharp.Completion.Presentation/IMyRoslynCompletionDataProvider.cs
@@ -2,7 +2,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Completion;
 using Microsoft.VisualStudio.Text;
 
-namespace MonoDevelop.Ide.CodeCompletion
+namespace MonoDevelop.CSharp.Completion.Presentation
 {
 	public interface IMyRoslynCompletionDataProvider
 	{
