+++ main/src/addins/CSharpBinding/MonoDevelop.CSharp.Completion.Presentation/MonoDevelopFormattingRuleFactoryServiceFactory.cs
@@ -8,7 +8,7 @@ using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.VisualStudio.Platform
+namespace MonoDevelop.CSharp.Completion.Presentation
 {
 	[ExportWorkspaceServiceFactory (typeof (IHostDependentFormattingRuleFactoryService), ServiceLayer.Host), Shared]
 	internal sealed class MonoDevelopFormattingRuleFactoryServiceFactory : IWorkspaceServiceFactory
