+++ main/src/addins/CSharpBinding/MonoDevelop.CSharp.Completion.Presentation/RoslynCompletionPresenter.cs
@@ -17,7 +17,7 @@ using Microsoft.VisualStudio.Utilities;
 using MonoDevelop.Ide;
 using MonoDevelop.Ide.CodeCompletion;
 
-namespace MonoDevelop.Ide.CodeCompletion
+namespace MonoDevelop.CSharp.Completion.Presentation
 {
     public interface IContentTypeMetadata
     {
