+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.csproj
@@ -9625,7 +9625,6 @@
     <Compile Include="MonoDevelop.Ide.Projects\LanguageCellRenderer.cs" />
     <Compile Include="MonoDevelop.Components\RestartPanel.cs" />
     <Compile Include="MonoDevelop.Ide.TypeSystem\MonoDevelopSourceText.cs" />
-    <Compile Include="MonoDevelop.Ide.CodeCompletion\CompletionPresenterSession.cs" />
     <Compile Include="MonoDevelop.Ide.CodeCompletion\TaggedTextUtil.cs" />
     <Compile Include="MonoDevelop.Ide.CodeCompletion\SignatureHelpParameterHintingData.cs" />
     <Compile Include="MonoDevelop.Ide.CodeCompletion\RoslynCompletionData.cs" />
